rename: 문의관련 컴포넌트 파일 이동 (#310)
