+++ src/components/page/mypage/inquires-detail/Question.tsx
