+++ src/components/page/mypage/inquires-detail/TraderAnswerView.tsx
