+++ src/components/page/mypage/inquires-edit/StrategyInfo.tsx
