+++ src/pages/mypage/investor/MyInquiresDetailPage.tsx
@@ -4,8 +4,8 @@ import { useNavigate, useParams } from 'react-router-dom';
 import Loader from '@/components/common/Loading';
 import Modal from '@/components/common/Modal';
 import Toast from '@/components/common/Toast';
-import Answer from '@/components/page/mypage-investor/inquires-detail/Answer';
-import Question from '@/components/page/mypage-investor/inquires-detail/Question';
+import Answer from '@/components/page/mypage/inquires-detail/Answer';
+import Question from '@/components/page/mypage/inquires-detail/Question';
 import { ROUTES } from '@/constants/routes';
 import { useDeleteMyInquiry } from '@/hooks/mutations/useMyInquiryMutations';
 import useFetchInquiryDetail from '@/hooks/queries/useFetchInquiryDetail';
