+++ src/pages/mypage/investor/MyInquiresEditPage.tsx
@@ -7,8 +7,8 @@ import Button from '@/components/common/Button';
 import Loader from '@/components/common/Loading';
 import Modal from '@/components/common/Modal';
 import Toast from '@/components/common/Toast';
-import InquiresInput from '@/components/page/mypage-investor/inquires-edit/InquiresInput';
-import StrategyInfo from '@/components/page/mypage-investor/inquires-edit/StrategyInfo';
+import InquiresInput from '@/components/page/mypage/inquires-edit/InquiresInput';
+import StrategyInfo from '@/components/page/mypage/inquires-edit/StrategyInfo';
 import { ROUTES } from '@/constants/routes';
 import { useUpdateMyInquiry } from '@/hooks/mutations/useMyInquiryMutations';
 import useFetchInquiryDetail from '@/hooks/queries/useFetchInquiryDetail';
